+++ motoman_driver/src/industrial_robot_client/joint_trajectory_interface.cpp
@@ -665,28 +665,28 @@ bool JointTrajectoryInterface::stopMotionCB(
 
 bool JointTrajectoryInterface::is_valid(const trajectory_msgs::JointTrajectory &traj)
 {
-  for (int i = 0; i < traj.points.size(); ++i)
+  for (size_t i = 0; i < traj.points.size(); ++i)
   {
     const trajectory_msgs::JointTrajectoryPoint &pt = traj.points[i];
 
     // check for non-empty positions
     if (pt.positions.empty())
-      ROS_ERROR_RETURN(false, "Validation failed: Missing position data for trajectory pt %d", i);
+      ROS_ERROR_RETURN(false, "Validation failed: Missing position data for trajectory pt %lu", i);
 
     // check for joint velocity limits
-    for (int j = 0; j < pt.velocities.size(); ++j)
+    for (size_t j = 0; j < pt.velocities.size(); ++j)
     {
       std::map<std::string, double>::iterator max_vel = joint_vel_limits_.find(traj.joint_names[j]);
       if (max_vel == joint_vel_limits_.end()) continue;  // no velocity-checking if limit not defined
 
       if (std::abs(pt.velocities[j]) > max_vel->second)
-        ROS_ERROR_RETURN(false, "Validation failed: Max velocity exceeded for trajectory pt %d, joint '%s'", i,
+        ROS_ERROR_RETURN(false, "Validation failed: Max velocity exceeded for trajectory pt %lu, joint '%s'", i,
                          traj.joint_names[j].c_str());
     }
 
     // check for valid timestamp
     if ((i > 0) && (pt.time_from_start.toSec() == 0))
-      ROS_ERROR_RETURN(false, "Validation failed: Missing valid timestamp data for trajectory pt %d", i);
+      ROS_ERROR_RETURN(false, "Validation failed: Missing valid timestamp data for trajectory pt %lu", i);
   }
 
   return true;
@@ -694,7 +694,7 @@ bool JointTrajectoryInterface::is_valid(const trajectory_msgs::JointTrajectory &
 
 bool JointTrajectoryInterface::is_valid(const motoman_msgs::DynamicJointTrajectory &traj)
 {
-  for (int i = 0; i < traj.points.size(); ++i)
+  for (size_t i = 0; i < traj.points.size(); ++i)
   {
     for (int gr = 0; gr < traj.points[i].num_groups; gr++)
     {
@@ -702,21 +702,21 @@ bool JointTrajectoryInterface::is_valid(const motoman_msgs::DynamicJointTrajecto
 
       // check for non-empty positions
       if (pt.positions.empty())
-        ROS_ERROR_RETURN(false, "Validation failed: Missing position data for trajectory pt %d", i);
+        ROS_ERROR_RETURN(false, "Validation failed: Missing position data for trajectory pt %lu", i);
       // check for joint velocity limits
-      for (int j = 0; j < pt.velocities.size(); ++j)
+      for (size_t j = 0; j < pt.velocities.size(); ++j)
       {
         std::map<std::string, double>::iterator max_vel = joint_vel_limits_.find(traj.joint_names[j]);
         if (max_vel == joint_vel_limits_.end()) continue;  // no velocity-checking if limit not defined
 
         if (std::abs(pt.velocities[j]) > max_vel->second)
-          ROS_ERROR_RETURN(false, "Validation failed: Max velocity exceeded for trajectory pt %d, joint '%s'", i,
+          ROS_ERROR_RETURN(false, "Validation failed: Max velocity exceeded for trajectory pt %lu, joint '%s'", i,
                            traj.joint_names[j].c_str());
       }
 
       // check for valid timestamp
       if ((i > 0) && (pt.time_from_start.toSec() == 0))
-        ROS_ERROR_RETURN(false, "Validation failed: Missing valid timestamp data for trajectory pt %d", i);
+        ROS_ERROR_RETURN(false, "Validation failed: Missing valid timestamp data for trajectory pt %lu", i);
     }
   }
   return true;
